Keep DeterministicEnvironment2 near its parent class
